Stop wrapping BadArgument in ConversionFailure
This is breaking but it's best to resolve it like this.

Most (but not all) of the functionality of ConversionFailure
can be replicated with Context.current_parameter.

It is (currently) impossible to get the equivalent of
`ConversionFailure.argument` from d.py alone.
If this turns out to be useful, we might try to implement on our own
though it would be preferred to have this directly in d.py.